\name{auscm-package}
\alias{auscm-package}
\alias{auscm}
\docType{package}

\title{
Contact Matrices for Australia
}

\description{
    A pure data package providing contact matrices and some other assorted data used for modelling
    COVID-19 interventions in Australia.
}

\details{
    These data are intended for use with the (Github) package \emph{sleepingwell/abcovid}. They are
    published via this package because they are released under creative commons licenses, which are
    not compatible with the Apache V2 licence used by that package.

    It is unlikely that the data will be useful outside of that package and we recommend that you go
    to the original source for the data. Please see the references provided below for how to do
    that.
}

\author{
Maintainer: Simon Knapp <simon.knapp@data61.csiro.au>
}

\references{
\insertRef{10.1371/journal.pcbi.1005697}{auscm}

\insertRef{10.1371/journal.pmed.0050074}{auscm}
}

\keyword{package}

\examples{
help(AusData)
}
