\name{auscm-package}
\alias{auscm-package}
\alias{auscm}
\docType{package}

\title{
Contact Matricies for Australia
}

\description{
A pure data package providing contact matrices for Australia.
}

%%\details{
%%}

\author{
Maintainer: Simon Knapp <simon.knapp@data61.csiro.au>
}

\references{
\insertRef{10.1371/journal.pcbi.1005697}{auscm}

\insertRef{10.1371/journal.pmed.0050074}{auscm}
}

\keyword{package}

%%\examples{
%%}
